Add alert diaolog for deletion

+++ components/delete-button.tsx
@@ -3,17 +3,47 @@
 import { deleteExpense } from "@/lib/actions";
 import { Button } from "./ui/button";
 import { Trash2 } from "lucide-react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 export function DeleteButton({ id }: { id: number }) {
   return (
-    <Button
-      onClick={async () => {
-        await deleteExpense(id);
-      }}
-      className="ml-auto"
-      variant="destructive"
-    >
-      <Trash2 />
-    </Button>
+    <AlertDialog>
+      <AlertDialogTrigger
+        render={
+          <Button className="ml-auto" variant="destructive">
+            <Trash2 />
+          </Button>
+        }
+      ></AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. This will permanently delete your
+            expense from our servers.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <AlertDialogAction
+            onClick={async () => {
+              await deleteExpense(id);
+            }}
+          >
+            Delete
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 }
